Secure the application

+++ app/api/issues/[id]/route.ts
@@ -1,11 +1,16 @@
+import authOptions from '@/app/auth/authOptions'
 import { issueSchema } from '@/app/validationSchemas'
 import prisma from '@/prisma/client'
+import { getServerSession } from 'next-auth'
 import { NextRequest, NextResponse } from 'next/server'
 
 export async function PATCH(
   request: NextRequest,
   { params }: { params: { id: string } }
 ) {
+  const session = await getServerSession(authOptions)
+  if (!session) return NextResponse.json({}, { status: 401 })
+
   const body = await request.json()
   const validation = issueSchema.safeParse(body)
   if (!validation.success)
@@ -24,6 +29,7 @@ export async function PATCH(
       description: body.description,
     },
   })
+
   return NextResponse.json(updatedIssue)
 }
 
@@ -31,13 +37,19 @@ export async function DELETE(
   request: NextRequest,
   { params }: { params: { id: string } }
 ) {
+  const session = await getServerSession(authOptions)
+  if (!session) return NextResponse.json({}, { status: 401 })
+
   const issue = await prisma.issue.findUnique({
     where: { id: parseInt(params.id) },
   })
+
   if (!issue)
     return NextResponse.json({ error: 'Invalid issue' }, { status: 404 })
+
   await prisma.issue.delete({
     where: { id: issue.id },
   })
+
   return NextResponse.json({})
 }
